fastapi/pyproject.toml
[tool.poetry]
name = "django-fastapi-template-app"
version = "0.1.0"
description = ""
authors = ["kathmandu777"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
fastapi = "^0.78.0"
pydantic = "^1.9.1"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
whitenoise = "^6.2.0"
psycopg2 = "^2.9.3"
gunicorn = "^20.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
twilio = "^7.9.3"
Pillow = "^9.1.1"
pytz = "^2022.1"

[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
